requirements.txt
fire
simonspeckciphers
pyotp
stoyled
clipboard
